@@ -3,7 +3,8 @@
     "description": "PHP wrapper for Cronofy's unified calendar API",
     "version": "v1.0.0",
     "require": {
-        "php": "^7.1"
+        "php": "^7.2",
+        "friendsofphp/php-cs-fixer": "^2.16"
     },
     "autoload": {
         "files": [
@@ -11,6 +12,11 @@
         ],
         "psr-4": {
             "Cronofy\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Cronofy\\Tests\\": "tests/"
         }
     },
     "keywords": [
@@ -25,7 +31,16 @@
     "minimum-stability": "beta",
     "license": "MIT",
     "require-dev": {
-        "phpunit/phpunit": "^5.7",
-        "squizlabs/php_codesniffer": "3.*"
+        "squizlabs/php_codesniffer": "3.*",
+        "phpunit/phpunit": "^8.0",
+        "phpstan/phpstan": "^0.12.14",
+        "rector/rector": "^0.7.5"
+    },
+    "scripts": {
+        "test": "vendor/bin/phpunit --bootstrap vendor/autoload.php --configuration phpunit.xml",
+        "phpstan": "vendor/bin/phpstan analyse --level=5 src/ tests/",
+        "cs": "vendor/bin/php-cs-fixer fix -v --dry-run --diff --allow-risky=yes",
+        "fixcs": "vendor/bin/php-cs-fixer fix -v --allow-risky=yes",
+        "rector": "vendor/bin/rector process --config rector.yaml"
     }
 }
