@@ -1,36 +1,37 @@
 <?php
+
 namespace Cronofy\Tests;
 
+use Cronofy\Cronofy;
 use Cronofy\Http\HttpRequest;
 use PHPUnit\Framework\TestCase;
-use Cronofy\Cronofy;
 
 class RulesTest extends TestCase
 {
-    public function testGetAvailabilityRule()
+    public function testGetAvailabilityRule(): void
     {
         $expected_rule = [
-            "availability_rule" => [
-                "availability_rule_id" => "default",
-                "tzid" => "Etc/UTC",
-                "weekly_periods" => [
+            'availability_rule' => [
+                'availability_rule_id' => 'default',
+                'tzid' => 'Etc/UTC',
+                'weekly_periods' => [
                     [
-                        "day" => "monday",
-                        "start_time" => "09:00",
-                        "end_time" => "13:00"
+                        'day' => 'monday',
+                        'start_time' => '09:00',
+                        'end_time' => '13:00',
                     ],
                     [
-                        "day" => "monday",
-                        "start_time" => "14:00",
-                        "end_time" => "17:00"
+                        'day' => 'monday',
+                        'start_time' => '14:00',
+                        'end_time' => '17:00',
                     ],
                     [
-                        "day" => "tuesday",
-                        "start_time" => "09:00",
-                        "end_time" => "17:00"
-                    ]
-                ]
-            ]
+                        'day' => 'tuesday',
+                        'start_time' => '09:00',
+                        'end_time' => '17:00',
+                    ],
+                ],
+            ],
         ];
         $http = $this->createMock(HttpRequest::class);
         $http->expects($this->once())
@@ -38,67 +39,67 @@ class RulesTest extends TestCase
             ->with(
                 $this->equalTo('https://api.cronofy.com/v1/availability_rules/default')
             )
-            ->will($this->returnValue([json_encode($expected_rule), 200]));
+            ->will($this->returnValue([\json_encode($expected_rule), 200]));
 
         $cronofy = new Cronofy([
-            "client_id" => "clientId",
-            "client_secret" => "clientSecret",
-            "access_token" => "accessToken",
-            "refresh_token" => "refreshToken",
-            "http_client" => $http,
+            'client_id' => 'clientId',
+            'client_secret' => 'clientSecret',
+            'access_token' => 'accessToken',
+            'refresh_token' => 'refreshToken',
+            'http_client' => $http,
         ]);
 
-        $response = $cronofy->getAvailabilityRule("default");
+        $response = $cronofy->getAvailabilityRule('default');
 
         $this->assertNotNull($response);
-        $this->assertEquals(3, count($response['availability_rule']));
-        $this->assertEquals("default", $response['availability_rule']['availability_rule_id']);
+        $this->assertCount(3, $response['availability_rule']);
+        $this->assertEquals('default', $response['availability_rule']['availability_rule_id']);
     }
 
-    public function testListAvailabilityRules()
+    public function testListAvailabilityRules(): void
     {
         $expected_rules = [
-            "availability_rules" => [
+            'availability_rules' => [
                 [
-                    "availability_rule_id" => "default",
-                    "calendar_ids" => ["cal_123"],
-                    "tzid" => "Etc/UTC",
-                    "weekly_periods" => [
+                    'availability_rule_id' => 'default',
+                    'calendar_ids' => ['cal_123'],
+                    'tzid' => 'Etc/UTC',
+                    'weekly_periods' => [
                         [
-                            "day" => "monday",
-                            "start_time" => "09:00",
-                            "end_time" => "13:00"
+                            'day' => 'monday',
+                            'start_time' => '09:00',
+                            'end_time' => '13:00',
                         ],
                         [
-                            "day" => "monday",
-                            "start_time" => "14:00",
-                            "end_time" => "17:00"
+                            'day' => 'monday',
+                            'start_time' => '14:00',
+                            'end_time' => '17:00',
                         ],
                         [
-                            "day" => "tuesday",
-                            "start_time" => "09:00",
-                            "end_time" => "17:00"
-                        ]
-                    ]
+                            'day' => 'tuesday',
+                            'start_time' => '09:00',
+                            'end_time' => '17:00',
+                        ],
+                    ],
                 ],
                 [
-                    "availability_rule_id" => "work_hours",
-                    "calendar_ids" => ["cal_321"],
-                    "tzid" => "Etc/UTC",
-                    "weekly_periods" => [
+                    'availability_rule_id' => 'work_hours',
+                    'calendar_ids' => ['cal_321'],
+                    'tzid' => 'Etc/UTC',
+                    'weekly_periods' => [
                         [
-                            "day" => "tuesday",
-                            "start_time" => "09:00",
-                            "end_time" => "17:00"
+                            'day' => 'tuesday',
+                            'start_time' => '09:00',
+                            'end_time' => '17:00',
                         ],
                         [
-                            "day" => "wednesday",
-                            "start_time" => "09:00",
-                            "end_time" => "17:00"
-                        ]
-                    ]
+                            'day' => 'wednesday',
+                            'start_time' => '09:00',
+                            'end_time' => '17:00',
+                        ],
+                    ],
                 ],
-            ]
+            ],
         ];
         $http = $this->createMock(HttpRequest::class);
         $http->expects($this->once())
@@ -106,28 +107,27 @@ class RulesTest extends TestCase
             ->with(
                 $this->equalTo('https://api.cronofy.com/v1/availability_rules')
             )
-            ->will($this->returnValue([json_encode($expected_rules), 200]));
+            ->will($this->returnValue([\json_encode($expected_rules), 200]));
 
         $cronofy = new Cronofy([
-            "client_id" => "clientId",
-            "client_secret" => "clientSecret",
-            "access_token" => "accessToken",
-            "refresh_token" => "refreshToken",
-            "http_client" => $http,
+            'client_id' => 'clientId',
+            'client_secret' => 'clientSecret',
+            'access_token' => 'accessToken',
+            'refresh_token' => 'refreshToken',
+            'http_client' => $http,
         ]);
 
         $response = $cronofy->listAvailabilityRules();
 
         $this->assertNotNull($response);
-        $this->assertEquals(2, count($response['availability_rules']));
-        $this->assertEquals(4, count($response['availability_rules'][0]));
-        $this->assertEquals("default", $response['availability_rules'][0]['availability_rule_id']);
-        $this->assertEquals("work_hours", $response['availability_rules'][1]['availability_rule_id']);
+        $this->assertCount(2, $response['availability_rules']);
+        $this->assertCount(4, $response['availability_rules'][0]);
+        $this->assertEquals('default', $response['availability_rules'][0]['availability_rule_id']);
+        $this->assertEquals('work_hours', $response['availability_rules'][1]['availability_rule_id']);
     }
 
-    public function testDeleteAvailabilityRules()
+    public function testDeleteAvailabilityRules(): void
     {
-
         $http = $this->createMock(HttpRequest::class);
         $http->expects($this->once())
             ->method('httpDelete')
@@ -137,53 +137,53 @@ class RulesTest extends TestCase
             ->will($this->returnValue(["{'foo': 'bar'}", 200]));
 
         $cronofy = new Cronofy([
-            "client_id" => "clientId",
-            "client_secret" => "clientSecret",
-            "access_token" => "accessToken",
-            "refresh_token" => "refreshToken",
-            "http_client" => $http,
+            'client_id' => 'clientId',
+            'client_secret' => 'clientSecret',
+            'access_token' => 'accessToken',
+            'refresh_token' => 'refreshToken',
+            'http_client' => $http,
         ]);
 
-        $actual = $cronofy->deleteAvailabilityRule("rule_123");
+        $actual = $cronofy->deleteAvailabilityRule('rule_123');
         $this->assertNotNull($actual);
     }
 
-    public function testCreateAvailabilityRule()
+    public function testCreateAvailabilityRule(): void
     {
         $expected_output = [
-            "availability_rule_id" => "default",
-            "calendar_ids" => ["cal_123"],
-            "tzid" => "America/Chicago",
-            "weekly_periods" => [
+            'availability_rule_id' => 'default',
+            'calendar_ids' => ['cal_123'],
+            'tzid' => 'America/Chicago',
+            'weekly_periods' => [
                 [
-                    "day" => "monday",
-                    "start_time" => "09:30",
-                    "end_time" => "12:30"
+                    'day' => 'monday',
+                    'start_time' => '09:30',
+                    'end_time' => '12:30',
                 ],
                 [
-                    "day" => "wednesday",
-                    "start_time" => "09:30",
-                    "end_time" => "12:30"
-                ]
-            ]
+                    'day' => 'wednesday',
+                    'start_time' => '09:30',
+                    'end_time' => '12:30',
+                ],
+            ],
         ];
 
         $params = [
-            "availability_rule_id" => "default",
-            "calendar_ids" => ["cal_123"],
-            "tzid" => "America/Chicago",
-            "weekly_periods" => [
+            'availability_rule_id' => 'default',
+            'calendar_ids' => ['cal_123'],
+            'tzid' => 'America/Chicago',
+            'weekly_periods' => [
                 [
-                    "day" => "monday",
-                    "start_time" => "09:30",
-                    "end_time" => "12:30"
+                    'day' => 'monday',
+                    'start_time' => '09:30',
+                    'end_time' => '12:30',
                 ],
                 [
-                    "day" => "wednesday",
-                    "start_time" => "09:30",
-                    "end_time" => "12:30"
-                ]
-            ]
+                    'day' => 'wednesday',
+                    'start_time' => '09:30',
+                    'end_time' => '12:30',
+                ],
+            ],
         ];
 
         $http = $this->createMock(HttpRequest::class);
@@ -193,21 +193,21 @@ class RulesTest extends TestCase
                 $this->equalTo('https://api.cronofy.com/v1/availability_rules'),
                 $this->equalTo($params)
             )
-            ->will($this->returnValue([json_encode($expected_output), 200]));
+            ->will($this->returnValue([\json_encode($expected_output), 200]));
 
         $cronofy = new Cronofy([
-            "client_id" => "clientId",
-            "client_secret" => "clientSecret",
-            "access_token" => "accessToken",
-            "refresh_token" => "refreshToken",
-            "http_client" => $http,
+            'client_id' => 'clientId',
+            'client_secret' => 'clientSecret',
+            'access_token' => 'accessToken',
+            'refresh_token' => 'refreshToken',
+            'http_client' => $http,
         ]);
 
         $response = $cronofy->createAvailabilityRule($params);
 
         $this->assertNotNull($response);
-        $this->assertEquals(4, count($response));
-        $this->assertEquals("default", $response['availability_rule_id']);
-        $this->assertEquals(2, count($response['weekly_periods']));
+        $this->assertCount(4, $response);
+        $this->assertEquals('default', $response['availability_rule_id']);
+        $this->assertCount(2, $response['weekly_periods']);
     }
 }
